refactor(subtitles): extract Scoring.Factor and Scoring.Reference to their own files
The project forbids nesting multiple modules in one file, since it can
cause cyclic dependencies and compilation errors. scoring.ex nested
Factor and Reference inside Scoring; split them into
scoring/factor.ex and scoring/reference.ex and alias both back in.
